refactor(neoscroll) rename to smooth
